pyproject.toml
[project]
name = "project-multi-agents-benchmark"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11,<3.13"
dependencies = [
    "absl-py>=2.2.2",
    "asteval>=0.9.31",
    "immutabledict>=4.2.1",
    "langchain>=0.2.25",
    "langchain-community>=0.3.19",
    "langchain-openai>=0.3.7",
    "langdetect>=1.0.9",
    "langgraph>=0.2.74",
    "langsmith>=0.3.11",
    "latex2sympy2-extended>=1.2.3",
    "matplotlib>=3.10.1",
    "math-verify[antlr4-13-2]>=0.7.0",
    "nest-asyncio>=1.6.0",
    "networkx>=3.4.2",
    "nltk>=3.9.1",
    "numpy>=1.26.4",
    "openai>=1.47.1",
    "pandas>=2.2.3",
    "playwright>=1.46.0",
    "psutil>=7.0.0",
    "pydantic>=2.8.2",
    "python-dotenv>=1.0.1",
    "seaborn>=0.13.2",
    "tavily-python>=0.3.0",
    "tree-sitter-python>=0.23.6",
    "tree-sitter>=0.24.0",
    "sentence-transformers",
    "gputil",
    "docx2markdown",
    "pypdf2",
    "langchain-mcp-adapters",
    "langchain-core",
    "aliyun-python-sdk-core",
    "requests>=2.31.0",
    "tika",
    "xmltodict",
    "aiofiles>=24.1.0",
    "rich>=14.0.0",
    "stopit",
    "tqdm>=4.64.1",
    "sympy",
    "torch",
    "transformers",
    "pyyaml>=6.0",
    "regex>=2023.6.3",
    "pydantic-core",
    "aiohttp>=3.8.5",
    "beautifulsoup4==4.12.2",
    "colorama==0.4.6",
    "easydict==1.10",
    "faiss-cpu==1.7.4",
    "flask==2.3.2",
    "flask-socketio==5.3.4",
    "importlib-metadata==6.8.0",
    "markdown==3.4.4",
    "pillow==10.3.0",
    "tenacity==8.2.2",
    "virtualenv==20.23.0",
    "werkzeug==3.0.3",
    "wikipedia-api==0.6.0",
    "duckduckgo-search>=8.1.1",
    "wolframalpha>=5.1.3",
    "langchain-experimental>=0.3.4",
]

[tool.ruff]
line-length = 120
target-version = "py311"
indent-width = 4

[tool.ruff.lint]
extend-select = ["C4", "SIM"]
ignore = ["F401"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
docstring-code-format = true

[[tool.uv.index]]
url = "https://mirrors.aliyun.com/pypi/simple"
default = true
